requirements.txt
praw==4.6.0
mysql-connector-python==8.0.5
